+++ RELEASE-NOTES.md
@@ -1,5 +1,8 @@
 ## RELEASE NOTES
 
+### Version 4.13.0-case-flags-tab-and-banner-message
+**EUI-4836** Add case flags tab to case details
+**EUI-4897** Add banner message to case details
 
 ### Version 4.12.1-EUI-5225-organisation-search
 **EUI-5225** Bug fix the organisation search functionality